Updated pre-commit setup



See merge request !4

+++ tools/setup.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
-cp dev/pre-commit .git/hooks/pre-commit
+cp tools/pre-commit .git/hooks/pre-commit
 chmod +x .git/hooks/pre-commit
-ln -s ../../../../drupal/coder/coder_sniffer/Drupal vendor/squizlabs/php_codesniffer/CodeSniffer/Standards/Drupal
+
+# Create symlink
+if [ ! -L vendor/squizlabs/php_codesniffer/CodeSniffer/Standards/Drupal/Drupal ]
+  then
+    ln -s ../../../../drupal/coder/coder_sniffer/Drupal vendor/squizlabs/php_codesniffer/CodeSniffer/Standards/Drupal
+fi
